fix race between distributed DDL and dropping Replicated db
