+++ src/Interpreters/InterpreterUpdateQuery.cpp
@@ -104,7 +104,7 @@ BlockIO InterpreterUpdateQuery::execute()
     {
         auto guard = DatabaseCatalog::instance().getDDLGuard(table_id.database_name, table_id.table_name, database.get());
         guard->releaseTableLock();
-        return database->tryEnqueueReplicatedDDL(query_ptr, getContext(), {});
+        return database->tryEnqueueReplicatedDDL(query_ptr, getContext(), {}, std::move(guard));
     }
 
     MutationCommands commands;
