+++ src/Interpreters/ReplicatedDatabaseQueryStatusSource.h
@@ -11,7 +11,7 @@ class ReplicatedDatabaseQueryStatusSource final : public DistributedQueryStatusS
 {
 public:
     ReplicatedDatabaseQueryStatusSource(
-        const String & zk_node_path, const String & zk_replicas_path, ContextPtr context_, const Strings & hosts_to_wait);
+        const String & zk_node_path, const String & zk_replicas_path, ContextPtr context_, const Strings & hosts_to_wait, DDLGuardPtr && database_guard_);
 
     String getName() const override { return "ReplicatedDatabaseQueryStatus"; }
 
@@ -26,6 +26,7 @@ protected:
 
 private:
     static Block getSampleBlock();
+    DDLGuardPtr database_guard;
 
 
 };
